requirements.txt
fastapi==0.111.0
d4kms==0.14.0
itsdangerous==2.2.0
starlette==0.37.2   
Authlib==1.3.1
sqlalchemy==2.0.31   
alembic==1.13.2
python-docx==1.1.2
docx2txt==0.8
fhir.resources==7.1.0
usdm==0.53.0
pytest==8.2.2
pytest-mock==3.14.0
beautifulsoup4==4.12.3
playwright==1.48.0
pytest-playwright==0.5.2
auth0-python==4.7.2
ruff==0.9.1
